+++ internal/conn/conn.go
@@ -420,8 +420,6 @@ func (c *conn) Invoke(
 		return xerrors.WithStackTrace(err)
 	}
 
-	invokeOpts := append([]grpc.CallOption(nil), opts...)
-	invokeOpts = append(invokeOpts, grpc.Trailer(&md))
 	opID, issues, err = invoke(
 		ctx,
 		method,
@@ -430,7 +428,7 @@ func (c *conn) Invoke(
 		cc,
 		c.Address(),
 		c.NodeID(),
-		invokeOpts...,
+		append(opts, grpc.Trailer(&md))...,
 	)
 
 	return err
@@ -487,12 +485,7 @@ func (c *conn) NewStream(
 		sentMark:   sentMark,
 	}
 
-	streamOpts := append([]grpc.CallOption(nil), opts...)
-	streamOpts = append(streamOpts,
-		grpc.Trailer(&s.trailer),
-		grpc.OnFinish(s.finish),
-	)
-	s.stream, err = cc.NewStream(grpcCtx, desc, method, streamOpts...)
+	s.stream, err = cc.NewStream(grpcCtx, desc, method, append(opts, grpc.OnFinish(s.finish))...)
 	if err != nil {
 		if xerrors.IsContextError(err) {
 			return nil, xerrors.WithStackTrace(err)